FIX :: 신청 화면 반응형 깨짐
