+++ src/components/ApplicationManagement/NoticeAside/NoticeAside.scss
@@ -101,7 +101,7 @@
   }
 }
 
-@media screen and (max-width: 590px) {
+@media screen and (max-width: 600px) {
   .notice-aside.responsive {
     width: 100vw;
   }
